agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stealth-agent"
version = "0.1.0"
description = "Agentic layer for converting product signals into repository PRs"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "anthropic>=0.42",
    "openai>=1.60",
    "pyjwt>=2.10",
    "structlog>=24.4",
    "httpx>=0.27",
]

[project.scripts]
stealth-agent = "stealth_agent.cli:main"
stealth-agent-serve = "stealth_agent.serve:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
pythonpath = ["."]
